+++ home/api.py
@@ -135,13 +135,73 @@ class OrderedContentSetViewSet(BaseAPIViewSet):
         "profile_fields",
         "pages",
     ]
-    known_query_parameters = BaseAPIViewSet.known_query_parameters.union(["page", "qa"])
+    known_query_parameters = BaseAPIViewSet.known_query_parameters.union(
+        ["page", "qa", "gender", "age", "relationship"]
+    )
     pagination_class = PageNumberPagination
     search_fields = ["name", "profile_fields"]
     filter_backends = (SearchFilter,)
 
+    def _filter_queryset_by_profile_fields(
+        self, ocs, gender_value, age_value, relationship_value
+    ):
+        pf_gender = ocs.profile_fields.blocks_by_name("gender")
+        pf_age = ocs.profile_fields.blocks_by_name("age")
+        pf_relationship = ocs.profile_fields.blocks_by_name("relationship")
+
+        pf_gender = pf_gender[0] if len(pf_gender) > 0 else None
+        pf_age = pf_age[0] if len(pf_age) > 0 else None
+        pf_relationship = pf_relationship[0] if len(pf_relationship) > 0 else None
+
+        if gender_value and age_value and relationship_value:
+            if (
+                pf_gender
+                and pf_gender.value == gender_value
+                and pf_age
+                and pf_age.value == age_value
+                and pf_relationship
+                and pf_relationship.value == relationship_value
+            ):
+                return ocs.id
+        elif gender_value and age_value:
+            if (
+                pf_gender
+                and pf_gender.value == gender_value
+                and pf_age
+                and pf_age.value == age_value
+            ):
+                return ocs.id
+        elif gender_value and relationship_value:
+            if (
+                pf_gender
+                and pf_gender.value == gender_value
+                and pf_relationship
+                and pf_relationship.value == relationship_value
+            ):
+                return ocs.id
+        elif age_value and relationship_value:
+            if (
+                pf_age
+                and pf_age.value == age_value
+                and pf_relationship
+                and pf_relationship.value == relationship_value
+            ):
+                return ocs.id
+        elif gender_value:
+            if pf_gender and pf_gender.value == gender_value:
+                return ocs.id
+        elif age_value:
+            if pf_age and pf_age.value == age_value:
+                return ocs.id
+        elif relationship_value:
+            if pf_relationship and pf_relationship.value == relationship_value:
+                return ocs.id
+
     def get_queryset(self):
         qa = self.request.query_params.get("qa")
+        gender = self.request.query_params.get("gender", "")
+        age = self.request.query_params.get("age", "")
+        relationship = self.request.query_params.get("relationship", "")
 
         if qa:
             # return the latest revision for each OrderedContentSet
@@ -155,9 +215,25 @@ class OrderedContentSetViewSet(BaseAPIViewSet):
                     ocs.profile_fields = latest_revision.profile_fields
 
         else:
-            queryset = OrderedContentSet.objects.filter(live=True).order_by(
-                "last_published_at"
-            )
+            if gender or age or relationship:
+                # it looks like you can't use advanced queries to filter on StreamFields
+                # so we have to do it like this.``
+                queryset = OrderedContentSet.objects.filter(
+                    live=True,
+                )
+                filter_ids = [
+                    self._filter_queryset_by_profile_fields(
+                        x, gender, age, relationship
+                    )
+                    for x in queryset
+                ]
+                queryset = OrderedContentSet.objects.filter(id__in=filter_ids).order_by(
+                    "last_published_at"
+                )
+            else:
+                queryset = OrderedContentSet.objects.filter(
+                    live=True,
+                ).order_by("last_published_at")
         return queryset
 
 
@@ -199,6 +275,7 @@ class AssessmentViewSet(BaseAPIViewSet):
         [
             "tag",
             "qa",
+            "page",
         ]
     )
     model = Assessment